compile-time tests to workaround rustc bug
